zxcvbn: Add self-contained python zxcvbn
This needed a bit of work to make it not dependent on any data files. A
very large amount of the program is a frequency list, which is simply a
LZMA compressed large JSON list. matching.py and scoring.py is also just
copy pasted into the file and the namespaces fixed.