i3: Don't autofocus new windows
This is done because while the default behaviour might work fine with
just one monitor, if I have my email client on a second monitor and open
a link, I don't want my web browser (Open on the other monitor) to be
focused, and cause any keys intended for the mail client to go towards
the web browser. (This is usually the case of me trying to delete an
email after I view a link in it, and the `d` key closes the tab in
qutebrowser).